src/requirements.txt
# API
fastapi==0.110.2
uvicorn[standard]==0.29.0

# AI
google-generativeai==0.5.2

# Code quality
black==24.4.1
isort==5.13.2

# Utilities
python-dotenv==1.0.1
opencv-python==4.9.0.80